use operation handler to delete objects in a thread safe way

+++ internal/api/object/handler.go
@@ -1,8 +1,10 @@
 package object
 
 import (
+	"errors"
 	"fmt"
 	"github.com/rstdm/glados/internal/api/object/file"
+	"github.com/rstdm/glados/internal/api/object/operationstate"
 	"go.uber.org/zap"
 	"mime/multipart"
 )
@@ -10,8 +12,9 @@ import (
 var ErrObjectAlreadyExists = file.ErrObjectAlreadyExists
 
 type Handler struct {
-	fileHandler *file.Handler
-	sugar       *zap.SugaredLogger
+	fileHandler           *file.Handler
+	operationStateHandler *operationstate.Handler
+	sugar                 *zap.SugaredLogger
 }
 
 func NewHandler(objectFolder string, sugar *zap.SugaredLogger) (*Handler, error) {
@@ -21,16 +24,57 @@ func NewHandler(objectFolder string, sugar *zap.SugaredLogger) (*Handler, error)
 		return nil, err
 	}
 
+	operationStateHandler := operationstate.NewHandler(sugar)
+
 	objectHandler := &Handler{
-		fileHandler: fileHandler,
-		sugar:       sugar,
+		fileHandler:           fileHandler,
+		operationStateHandler: operationStateHandler,
+		sugar:                 sugar,
 	}
 
 	return objectHandler, nil
 }
 
 func (h *Handler) DeleteObject(objectHash string) (didExist bool, err error) {
-	return h.fileHandler.DeleteObject(objectHash) // TODO improve
+	err = h.operationStateHandler.StartDeleting(objectHash, h.deleteCallback(objectHash))
+
+	if err != nil && errors.Is(err, operationstate.ErrDeletionMustBeDelayed) {
+		// pretend that the object has been deleted
+		// to the outside world it looks as if the object is gone (e.g. it can no longer be read)
+		return true, nil
+	}
+
+	// this defer must not be called if the deletion has been delayed
+	defer h.operationStateHandler.DoneDeleting(objectHash)
+
+	if err == nil {
+		return h.fileHandler.DeleteObject(objectHash)
+	}
+
+	// there was an error
+
+	if errors.Is(err, operationstate.ErrOperationNotAllowed) {
+		// pretend that the object doesn't exist
+		// the object is either in the process of being created or it has already been marked for deletion
+		return false, nil
+	}
+
+	//unexpected error
+	err = fmt.Errorf("mark object state as deleting: %w", err)
+	return false, err
+}
+
+func (h *Handler) deleteCallback(objectHash string) func() {
+	return func() {
+		defer h.operationStateHandler.DoneDeleting(objectHash)
+
+		if _, err := h.fileHandler.DeleteObject(objectHash); err != nil {
+			h.sugar.Errorw("Failed to delete object",
+				"err", err,
+				"objectHash", objectHash,
+			)
+		}
+	}
 }
 
 func (h *Handler) GetObjectPath(objectHash string) (string, error) {
